Add getPools/[blockchainId]/[registryId] endpoint to retrieve all pools info
