+++ pages/api/getFactoryAPYs-optimism.js
@@ -4,7 +4,6 @@ import { IS_DEV } from 'constants/AppConstants';
 
 import configs from '../../constants/configs';
 import { fn } from '../../utils/api';
-import { getOpMulticall } from '../../utils/getters';
 import registryAbi from '../../constants/abis/factory_registry.json';
 import multicallAbi from '../../constants/abis/multicall.json';
 import erc20Abi from '../../constants/abis/erc20.json';
@@ -19,7 +18,7 @@ export default fn(async (query) => {
     const version = 2
 
     let registryAddress = '0x2db0E83599a91b508Ac268a6197b8B14F5e72840'
-    let multicallAddress = getOpMulticall()
+    let multicallAddress = config.multicallAddress;
   	let registry = new web3.eth.Contract(registryAbi, registryAddress);
   	let multicall = new web3.eth.Contract(multicallAbi, multicallAddress)
     let res = await (await fetch(`${BASE_API_DOMAIN}/api/getFactoryV2Pools/optimism`)).json()
