+++ utils/getters.js
@@ -1,7 +1,8 @@
 import memoize from 'memoizee';
 import Web3 from 'web3';
 import WEB3_CONSTANTS from 'constants/Web3';
-import addressGetterAbi from '../constants/abis/address_getter.json';
+import configs from 'constants/configs';
+import addressGetterAbi from 'constants/abis/address_getter.json';
 const web3 = new Web3(WEB3_CONSTANTS.RPC_URL);
 const addressGetter = '0x0000000022d53366457f9d5e68ec105046fc4383'
 const multiCall = '0xeefBa1e63905eF1D7ACbA5a8513c70307C1cE441'
@@ -18,7 +19,10 @@ const feeDistributor = '0xA464e6DCda8AC41e03616F95f4BC98a13b8922Dc'
 
 
 
-const getRegistry = memoize(async () => {
+const getRegistry = memoize(async ({ blockchainId }) => {
+  if (typeof blockchainId === 'undefined') blockchainId = 'ethereum'; // eslint-disable-line no-param-reassign
+
+  const web3 = new Web3(configs[blockchainId].rpcUrl);
   const contract = new web3.eth.Contract(addressGetterAbi, addressGetter);
   return contract.methods.get_registry().call();
 }, {
@@ -94,6 +98,5 @@ export {
   getFactoryRegistry,
   getFeeDistributor,
   getMultiCall,
-  getRegistry,
-  getOpMulticall
+  getRegistry
 };
